+++ examples/examples_nodejs_test.go
@@ -1,0 +1,30 @@
+// Copyright 2016-2017, Pulumi Corporation.  All rights reserved.
+// +build nodejs all
+
+package examples
+
+import (
+	"testing"
+
+	"github.com/pulumi/pulumi/pkg/v2/testing/integration"
+)
+
+func TestSimple(t *testing.T) {
+	test := getJSBaseOptions(t).
+		With(integration.ProgramTestOptions{
+			Dir:           path.Join(getCwd(t), "simple"),
+		})
+
+	integration.ProgramTest(t, &test)
+}
+
+func getJSBaseOptions(t *testing.T) integration.ProgramTestOptions {
+	base := getBaseOptions()
+	baseJS := base.With(integration.ProgramTestOptions{
+		Dependencies: []string{
+			"@pulumi/random",
+		},
+	})
+
+	return baseJS
+}
